Create initial draft of the library.
This includes unit tests for all the major functions. Usage of the library
should be pretty simple, just guard your commit hook code with
with MessagePreservation():
  ...
And in your prepare-git-commit hook use get_cached_message() to retrieve
whatever was saved and everything should work.

I still have yet to test this with an actual hook, however.